requirements.txt
feedparser-6.0.12 
sgmllib3k-1.0.0